+++ src/commands/doctor.ts
@@ -3405,7 +3405,7 @@ export async function computeConversationFactsBacklogCheck(
     // list is the command's own ALLOWED_TYPES (not a hand-copied literal)
     // so this check never drifts out of sync with what the backfill
     // actually processes.
-    const { ALLOWED_TYPES } = await import('./extract-conversation-facts.ts');
+    const { ALLOWED_TYPES } = await import('../core/conversation-facts-types.ts');
     const typesRaw = await engine.getConfig(
       'cycle.conversation_facts_backfill.types',
     );
@@ -5788,7 +5788,7 @@ export async function buildChecks(
     try {
       const { readConversationBodyForParsing } = await import('../core/conversation-parser/body.ts');
       const { parseConversation } = await import('../core/conversation-parser/parse.ts');
-      const { ALLOWED_TYPES: allowedTypes } = await import('./extract-conversation-facts.ts');
+      const { ALLOWED_TYPES: allowedTypes } = await import('../core/conversation-facts-types.ts');
       // PageFilters supports singular `type` only; iterate the allowed types
       // and cap at ~50/each to land at ~200 total max.
       const sample: import('../core/types.ts').Page[] = [];
