+++ src/commands/sources.ts
@@ -1186,7 +1186,7 @@ async function runAudit(engine: BrainEngine, args: string[]): Promise<void> {
   // (~2000 in + 500 out tokens at $3/MTok in + $15/MTok out ≈ $0.013).
   // Reuses the command's own ALLOWED_TYPES (not a hand-copied literal) so
   // this estimator never under-reports pages the backfill actually processes.
-  const { ALLOWED_TYPES: FACTS_BACKFILL_ALLOWED } = await import('./extract-conversation-facts.ts');
+  const { ALLOWED_TYPES: FACTS_BACKFILL_ALLOWED } = await import('../core/conversation-facts-types.ts');
   const FACTS_BACKFILL_CHARS_PER_SEGMENT = 6500; // matches SEGMENT_TEXT_CHAR_LIMIT
   const FACTS_BACKFILL_USD_PER_SEGMENT = 0.013;
   let factsBackfillPages = 0;
